FastAPI Version/requirements.txt
fastapi>=0.70.0,<0.71.0
uvicorn[standard]>=0.15.0,<0.16.0
websocket-client>=1.2.1,<1.3.0
streamlit>=1.0.0,<1.1.0
websockets>=10.0,<11.0